refactor: move deviceType to configuration state
