+++ src/app/core/store/configuration/configuration.effects.ts
@@ -4,6 +4,7 @@ import { Actions, Effect, ROOT_EFFECTS_INIT, ofType } from '@ngrx/effects';
 import { routerNavigationAction } from '@ngrx/router-store';
 import { Store, select } from '@ngrx/store';
 import { TranslateService } from '@ngx-translate/core';
+import { defer, fromEvent, iif, merge } from 'rxjs';
 import {
   concatMap,
   distinctUntilChanged,
@@ -16,8 +17,10 @@ import {
   withLatestFrom,
 } from 'rxjs/operators';
 
+import { LARGE_BREAKPOINT_WIDTH, MEDIUM_BREAKPOINT_WIDTH } from 'ish-core/configurations/injection-keys';
+import { DeviceType } from 'ish-core/models/viewtype/viewtype.types';
 import { ConfigurationService } from 'ish-core/services/configuration/configuration.service';
-import { mapErrorToAction, mapToProperty, whenFalsy, whenTruthy } from 'ish-core/utils/operators';
+import { distinctCompareWith, mapErrorToAction, mapToProperty, whenFalsy, whenTruthy } from 'ish-core/utils/operators';
 import { StatePropertiesService } from 'ish-core/utils/state-transfer/state-properties.service';
 
 import {
@@ -27,7 +30,7 @@ import {
   LoadServerConfigFail,
   SetGTMToken,
 } from './configuration.actions';
-import { getCurrentLocale, isServerConfigurationLoaded } from './configuration.selectors';
+import { getCurrentLocale, getDeviceType, isServerConfigurationLoaded } from './configuration.selectors';
 
 @Injectable()
 export class ConfigurationEffects {
@@ -38,7 +41,9 @@ export class ConfigurationEffects {
     private translateService: TranslateService,
     private stateProperties: StatePropertiesService,
     @Inject(PLATFORM_ID) private platformId: string,
-    private appRef: ApplicationRef
+    private appRef: ApplicationRef,
+    @Inject(MEDIUM_BREAKPOINT_WIDTH) private mediumBreakpointWidth: number,
+    @Inject(LARGE_BREAKPOINT_WIDTH) private largeBreakpointWidth: number
   ) {}
 
   @Effect({ dispatch: false })
@@ -109,5 +114,25 @@ export class ConfigurationEffects {
     map(([, gtmToken]) => gtmToken),
     whenTruthy(),
     map(gtmToken => new SetGTMToken({ gtmToken }))
+  );
+
+  @Effect()
+  setDeviceType$ = iif(
+    () => isPlatformBrowser(this.platformId),
+    defer(() =>
+      merge(this.actions$.pipe(ofType(ROOT_EFFECTS_INIT)), fromEvent(window, 'resize')).pipe(
+        map(() => {
+          if (window.innerWidth < this.mediumBreakpointWidth) {
+            return 'mobile';
+          } else if (window.innerWidth < this.largeBreakpointWidth) {
+            return 'tablet';
+          } else {
+            return 'pc';
+          }
+        }),
+        distinctCompareWith(this.store.pipe(select(getDeviceType))),
+        map((deviceType: DeviceType) => new ApplyConfiguration({ _deviceType: deviceType }))
+      )
+    )
   );
 }
