+++ src/app/core/store/viewconf/viewconf.actions.ts
@@ -1,11 +1,9 @@
 import { Action } from '@ngrx/store';
 
 import { BreadcrumbItem } from 'ish-core/models/breadcrumb-item/breadcrumb-item.interface';
-import { DeviceType } from 'ish-core/models/viewtype/viewtype.types';
 
 export enum ViewconfActionTypes {
   SetBreadcrumbData = '[Viewconf Internal] Set Breadcrumb Data',
-  SetDeviceType = '[Viewconf Internal] Set Device Type',
   SetStickyHeader = '[Viewconf Internal] Set Sticky Header',
 }
 
@@ -14,14 +12,9 @@ export class SetBreadcrumbData implements Action {
   constructor(public payload: { breadcrumbData: BreadcrumbItem[] }) {}
 }
 
-export class SetDeviceType implements Action {
-  readonly type = ViewconfActionTypes.SetDeviceType;
-  constructor(public payload: { deviceType: DeviceType }) {}
-}
-
 export class SetStickyHeader implements Action {
   readonly type = ViewconfActionTypes.SetStickyHeader;
   constructor(public payload: { sticky: boolean }) {}
 }
 
-export type ViewconfActions = SetBreadcrumbData | SetDeviceType | SetStickyHeader;
+export type ViewconfActions = SetBreadcrumbData | SetStickyHeader;
