+++ src/app/core/store/viewconf/viewconf.effects.ts
@@ -1,47 +1,19 @@
 import { isPlatformBrowser } from '@angular/common';
 import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
-import { Actions, Effect, ROOT_EFFECTS_INIT, ofType } from '@ngrx/effects';
+import { Effect } from '@ngrx/effects';
 import { Store, select } from '@ngrx/store';
-import { defer, fromEvent, iif, merge } from 'rxjs';
+import { defer, fromEvent, iif } from 'rxjs';
 import { distinctUntilChanged, map } from 'rxjs/operators';
 
-import { LARGE_BREAKPOINT_WIDTH, MEDIUM_BREAKPOINT_WIDTH } from 'ish-core/configurations/injection-keys';
-import { DeviceType } from 'ish-core/models/viewtype/viewtype.types';
 import { selectRouteData } from 'ish-core/store/router';
 import { distinctCompareWith } from 'ish-core/utils/operators';
 
-import { SetBreadcrumbData, SetDeviceType, SetStickyHeader } from './viewconf.actions';
-import { getBreadcrumbData, getHeaderType } from './viewconf.selectors';
+import { SetBreadcrumbData, SetStickyHeader } from './viewconf.actions';
+import { getBreadcrumbData } from './viewconf.selectors';
 
 @Injectable()
 export class ViewconfEffects {
-  constructor(
-    private actions$: Actions,
-    private store: Store<{}>,
-    @Inject(PLATFORM_ID) private platformId: string,
-    @Inject(MEDIUM_BREAKPOINT_WIDTH) private mediumBreakpointWidth: number,
-    @Inject(LARGE_BREAKPOINT_WIDTH) private largeBreakpointWidth: number
-  ) {}
-
-  @Effect()
-  setDeviceType$ = iif(
-    () => isPlatformBrowser(this.platformId),
-    defer(() =>
-      merge(this.actions$.pipe(ofType(ROOT_EFFECTS_INIT)), fromEvent(window, 'resize')).pipe(
-        map(() => {
-          if (window.innerWidth < this.mediumBreakpointWidth) {
-            return 'mobile';
-          } else if (window.innerWidth < this.largeBreakpointWidth) {
-            return 'tablet';
-          } else {
-            return 'pc';
-          }
-        }),
-        distinctCompareWith(this.store.pipe(select(getHeaderType))),
-        map((deviceType: DeviceType) => new SetDeviceType({ deviceType }))
-      )
-    )
-  );
+  constructor(private store: Store<{}>, @Inject(PLATFORM_ID) private platformId: string) {}
 
   @Effect()
   toggleStickyHeader$ = iif(
