+++ src/app/core/store/viewconf/viewconf.integration.spec.ts
@@ -3,7 +3,6 @@ import { TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { LARGE_BREAKPOINT_WIDTH, MEDIUM_BREAKPOINT_WIDTH } from 'ish-core/configurations/injection-keys';
 import { TestStore, ngrxTesting } from 'ish-core/utils/dev/ngrx-testing';
 
 import { ViewconfEffects } from './viewconf.effects';
@@ -38,10 +37,6 @@ describe('Viewconf Integration', () => {
           },
           { path: '**', component: DummyComponent },
         ]),
-      ],
-      providers: [
-        { provide: MEDIUM_BREAKPOINT_WIDTH, useValue: 768 },
-        { provide: LARGE_BREAKPOINT_WIDTH, useValue: 992 },
       ],
     });
