+++ src/app/core/store/viewconf/viewconf.reducer.ts
@@ -1,19 +1,15 @@
 import { BreadcrumbItem } from 'ish-core/models/breadcrumb-item/breadcrumb-item.interface';
-import { DeviceType } from 'ish-core/models/viewtype/viewtype.types';
 
 import { ViewconfActionTypes, ViewconfActions } from './viewconf.actions';
 
 export interface ViewconfState {
   breadcrumbData: BreadcrumbItem[];
   stickyHeader: boolean;
-  // not synced via state transfer
-  _deviceType: DeviceType;
 }
 
 export const initialState: ViewconfState = {
   breadcrumbData: [],
   stickyHeader: false,
-  _deviceType: 'mobile',
 };
 
 export function viewconfReducer(state: ViewconfState = initialState, action: ViewconfActions): ViewconfState {
@@ -22,11 +18,6 @@ export function viewconfReducer(state: ViewconfState = initialState, action: Vie
       return {
         ...state,
         breadcrumbData: action.payload.breadcrumbData,
-      };
-    case ViewconfActionTypes.SetDeviceType:
-      return {
-        ...state,
-        _deviceType: action.payload.deviceType,
       };
     case ViewconfActionTypes.SetStickyHeader:
       return {
